requirements.txt
transformers
torch
faiss-cpu
spacy
pandas
matplotlib
wordcloud
click